remove more unused code
