+++ pkg/api/util_test.go
@@ -6,28 +6,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
 	resource "k8s.io/apimachinery/pkg/api/resource"
-
-	"github.com/armadaproject/armada/internal/common/types"
-)
-
-const (
-	PriorityClassEmpty = ""
-	PriorityClass0     = "priority-0"
-	PriorityClass1     = "priority-1"
-	PriorityClass2     = "priority-2"
-	PriorityClass3     = "priority-3"
-)
-
-var (
-	TestPriorityClasses = map[string]types.PriorityClass{
-		PriorityClassEmpty: {Priority: 0, Preemptible: true},
-		PriorityClass0:     {Priority: 0, Preemptible: true},
-		PriorityClass1:     {Priority: 1, Preemptible: true},
-		PriorityClass2:     {Priority: 2, Preemptible: true},
-		PriorityClass3:     {Priority: 3, Preemptible: false},
-	}
-	TestDefaultPriorityClass = PriorityClass3
-	TestPriorities           = []int32{0, 1, 2, 3}
 )
 
 func TestSchedulingResourceRequirementsFromPodSpec(t *testing.T) {
@@ -261,51 +239,4 @@ func QuantityWithMilliValue(v int64) resource.Quantity {
 	q := resource.Quantity{}
 	q.SetMilli(v)
 	return q
-}
-
-func TestPriorityFromPodSpec(t *testing.T) {
-	tests := map[string]struct {
-		podSpec          *v1.PodSpec
-		expectedPriority int32
-		expectedOk       bool
-	}{
-		"nil podSpec": {
-			podSpec:          nil,
-			expectedPriority: 0,
-			expectedOk:       false,
-		},
-		"priority already set": {
-			podSpec: &v1.PodSpec{
-				Priority:          pointerFromValue(int32(1)),
-				PriorityClassName: PriorityClass2,
-			},
-			expectedPriority: 1,
-			expectedOk:       true,
-		},
-		"existing priorityClass": {
-			podSpec: &v1.PodSpec{
-				PriorityClassName: PriorityClass2,
-			},
-			expectedPriority: 2,
-			expectedOk:       true,
-		},
-		"non-existing priorityClass": {
-			podSpec: &v1.PodSpec{
-				PriorityClassName: "does not exist",
-			},
-			expectedPriority: 0,
-			expectedOk:       false,
-		},
-	}
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			p, ok := PriorityFromPodSpec(tc.podSpec, TestPriorityClasses)
-			assert.Equal(t, tc.expectedPriority, p)
-			assert.Equal(t, tc.expectedOk, ok)
-		})
-	}
-}
-
-func pointerFromValue[T any](v T) *T {
-	return &v
 }
